update the split-driver with current master

+++ example/CO_driver.c
@@ -77,7 +77,7 @@ CO_ReturnError_t CO_CANmodule_init(
 
     for(i=0U; i<rxSize; i++){
         rxArray[i].ident = 0U;
-        rxArray[i].mask = 0xFFFFFFFFU;
+        rxArray[i].mask = (uint16_t)0xFFFFFFFFU;
         rxArray[i].object = NULL;
         rxArray[i].CANrx_callback = NULL;
     }
